feat(auth): complete user registration, login, JWT auth and admin management
- Custom exception system with error codes (ErrorCode enum)
- JWT dual-token authentication (access + refresh token)
- Registration with password strength check and toggle switch
- Login with rate limiting, failure lockout, and IP tracking
- Token refresh with old token blacklisting
- Admin CRUD with permission checks and audit logging
- Spring Security config with role-based URL and method security
- CORS configuration for local development
- Global exception handler with error code to HTTP status mapping
- Memory cache service based on Caffeine for token blacklist and rate limiting
- Permission service with JDBC query and role-based admin menus
- Fixed type mismatches (Integer -> Long for roleId)

+++ luomiblog-backend/src/main/java/com/luomiblog/security/UserDetailsServiceImpl.java
@@ -32,7 +32,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
                         .orElseThrow(() -> new UsernameNotFoundException("用户不存在: " + usernameOrEmail)));
 
         String roleCode = "visitor";
-        Integer roleId = user.getRoleId();
+        Long roleId = user.getRoleId();
         if (roleId != null) {
             Role role = roleRepository.findById(roleId).orElse(null);
             if (role != null) {
